requirements.txt
django
django_rest_framework
django-rest-knox
django-filter
